midi_input: add static for nInstrument to avoid passing a negative number to pInstrList->get()

+++ src/core/src/IO/midi_input.cpp
@@ -275,7 +275,7 @@ void MidiInput::handleNoteOnMessage( const MidiMessage& msg )
 				return;
 			}
 			
-			pInstr = pInstrList->get( nInstrument );
+			pInstr = pInstrList->get( static_cast<uint>(nInstrument) );
 		}
 
 		/*
